hasSkeletonElements tests
